moved queues to own namespace
