+++ src/main.ts
@@ -1,6 +1,3 @@
-import { AsyncQueue } from './AsyncQueue'
-import { BoundedAsyncQueue } from './BoundedAsyncQueue'
-import { UnboundedQueue } from './UnboundedQueue'
 
 let queue = new BoundedAsyncQueue<string>(1);
 // let queue = new UnboundedQueue<string>();
