+++ src/queue/BlockingQueue.ts
@@ -1,0 +1,9 @@
+/**
+ * Generic interface for a blocking Queue.
+ */
+
+export abstract class BlockingQueue<T> {
+    abstract push (val: T) : void;
+    abstract pop () : Promise<T>;
+}
+
